UI for Weather APP

+++ app/src/main/java/com/example/weatherapp/retrofit/Sys.kt
@@ -18,9 +18,9 @@ data class Sys (
 
     @SerializedName("sunrise")
     @Expose
-    var sunrise: Int? = null,
+    var sunrise: Long? = null,
 
     @SerializedName("sunset")
     @Expose
-    var sunset: Int? = null
+    var sunset: Long? = null
 )
